+++ src/orquestra/sdk/_ray/_ray_logs.py
@@ -216,5 +216,6 @@ class DirectRayReader:
         return WorkflowLogs(
             per_task=logs_dict,
             env_setup=env_setup,
+            system=[],
             other=[],
         )